Rename myClient to MyClient
