+++ src/performance/PERF_CONTOUR_DATA_Mode2.test.ts
@@ -1,5 +1,5 @@
 import { CARTA } from "carta-protobuf";
-import { checkConnection, Stream} from './myClient';
+import { checkConnection, Stream} from './MyClient';
 import { MessageController } from "./MessageController";
 import config from "./config.json";
